+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
